% Generated by r6-generator-maven-plugin: do not edit by hand

\name{async_factory}
\alias{async_factory}
\title{no title}
\usage{async_factory()
}
\value{
an `RFuture` with methods `cancel()`, `isCancelled()`, `isDone()` and `get()`. 
The result of a `get()` call will be an 
R6 FactoryTest object: 

}
\description{
no description
}
\keyword{java api}


