% Generated by r6-generator-maven-plugin: do not edit by hand

\name{async_static_countdown}
\alias{async_static_countdown}
\title{no title}
\usage{async_static_countdown(
	label,
	rtimer
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{label}{label - (java expects a RCharacter)}
	\item{rtimer}{rtimer - (java expects a RInteger)}
	\if{html}{\out{</div>}}
}
\value{
an `RFuture` with methods `cancel()`, `isCancelled()`, `isDone()` and `get()`. 
The result of a `get()` call will be an 
RCharacter: 

}
\description{
no description
}
\keyword{java api}


