% Generated by r6-generator-maven-plugin: do not edit by hand

\name{BounceTest}

\alias{BounceTest}

\title{BounceTest}



\description{
missing description

Version: 0.0.0.9999

Generated: 2022-09-24T01:15:12.865691
}

\details{
	no details
}

\examples{
## -----------------------------------
## Construct new instance of BounceTest
## -----------------------------------
\dontrun{
J = testRapi::JavaApi$get()
# appropriate parameter values must be provided
instance = J$BounceTest$new()
}

## -----------------------------------
## Method `BounceTest$bounceNull(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceNull(x)
}

## -----------------------------------
## Method `BounceTest$bounceVoid()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceVoid()
}

## -----------------------------------
## Method `BounceTest$bounceString(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceString(x)
}

## -----------------------------------
## Method `BounceTest$bounceCharacter(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceCharacter(x)
}

## -----------------------------------
## Method `BounceTest$bounceCharacterVector(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceCharacterVector(x)
}

## -----------------------------------
## Method `BounceTest$bounceDouble(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceDouble(x)
}

## -----------------------------------
## Method `BounceTest$bounceNumeric(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceNumeric(x)
}

## -----------------------------------
## Method `BounceTest$bounceNumericVector(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceNumericVector(x)
}

## -----------------------------------
## Method `BounceTest$bounceInt(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceInt(x)
}

## -----------------------------------
## Method `BounceTest$bounceInteger(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceInteger(x)
}

## -----------------------------------
## Method `BounceTest$bounceIntegerVector(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceIntegerVector(x)
}

## -----------------------------------
## Method `BounceTest$bounceFactor(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceFactor(x)
}

## -----------------------------------
## Method `BounceTest$bounceFactorVector(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceFactorVector(x)
}

## -----------------------------------
## Method `BounceTest$bounceDate(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceDate(x)
}

## -----------------------------------
## Method `BounceTest$bounceDateVector(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceDateVector(x)
}

## -----------------------------------
## Method `BounceTest$bounceLogical(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceLogical(x)
}

## -----------------------------------
## Method `BounceTest$bounceLogicalVector(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceLogicalVector(x)
}

## -----------------------------------
## Method `BounceTest$bounceDataframe(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceDataframe(x)
}

## -----------------------------------
## Method `BounceTest$bounceList(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceList(x)
}

## -----------------------------------
## Method `BounceTest$bounceNamedList(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceNamedList(x)
}

## -----------------------------------
## Method `BounceTest$bounceArray(x)`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$bounceArray(x)
}
}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$BounceTest$new()}}
		}
	}
	\subsection{Class methods}{
		\itemize{
			\item \href{#method-bounceNull}{\code{instance$bounceNull(x)}}
			\item \href{#method-bounceVoid}{\code{instance$bounceVoid()}}
			\item \href{#method-bounceString}{\code{instance$bounceString(x)}}
			\item \href{#method-bounceCharacter}{\code{instance$bounceCharacter(x)}}
			\item \href{#method-bounceCharacterVector}{\code{instance$bounceCharacterVector(x)}}
			\item \href{#method-bounceDouble}{\code{instance$bounceDouble(x)}}
			\item \href{#method-bounceNumeric}{\code{instance$bounceNumeric(x)}}
			\item \href{#method-bounceNumericVector}{\code{instance$bounceNumericVector(x)}}
			\item \href{#method-bounceInt}{\code{instance$bounceInt(x)}}
			\item \href{#method-bounceInteger}{\code{instance$bounceInteger(x)}}
			\item \href{#method-bounceIntegerVector}{\code{instance$bounceIntegerVector(x)}}
			\item \href{#method-bounceFactor}{\code{instance$bounceFactor(x)}}
			\item \href{#method-bounceFactorVector}{\code{instance$bounceFactorVector(x)}}
			\item \href{#method-bounceDate}{\code{instance$bounceDate(x)}}
			\item \href{#method-bounceDateVector}{\code{instance$bounceDateVector(x)}}
			\item \href{#method-bounceLogical}{\code{instance$bounceLogical(x)}}
			\item \href{#method-bounceLogicalVector}{\code{instance$bounceLogicalVector(x)}}
			\item \href{#method-bounceDataframe}{\code{instance$bounceDataframe(x)}}
			\item \href{#method-bounceList}{\code{instance$bounceList(x)}}
			\item \href{#method-bounceNamedList}{\code{instance$bounceNamedList(x)}}
			\item \href{#method-bounceArray}{\code{instance$bounceArray(x)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Constructor \code{new()}}{
the default no-args constructor

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
the new R6 BounceTest object
		}

	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNull"></a>}}
	
	\subsection{Method \code{bounceNull()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceNull(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RNull)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RNull: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceVoid"></a>}}
	
	\subsection{Method \code{bounceVoid()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceVoid()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
void: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceString"></a>}}
	
	\subsection{Method \code{bounceString()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceString(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
String: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceCharacter"></a>}}
	
	\subsection{Method \code{bounceCharacter()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceCharacter(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RCharacter)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RCharacter: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceCharacterVector"></a>}}
	
	\subsection{Method \code{bounceCharacterVector()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceCharacterVector(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RCharacterVector)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RCharacterVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDouble"></a>}}
	
	\subsection{Method \code{bounceDouble()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceDouble(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
double: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNumeric"></a>}}
	
	\subsection{Method \code{bounceNumeric()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceNumeric(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RNumeric)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RNumeric: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNumericVector"></a>}}
	
	\subsection{Method \code{bounceNumericVector()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceNumericVector(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RNumericVector)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RNumericVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceInt"></a>}}
	
	\subsection{Method \code{bounceInt()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceInt(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
int: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceInteger"></a>}}
	
	\subsection{Method \code{bounceInteger()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceInteger(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RInteger)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RInteger: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceIntegerVector"></a>}}
	
	\subsection{Method \code{bounceIntegerVector()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceIntegerVector(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RIntegerVector)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RIntegerVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceFactor"></a>}}
	
	\subsection{Method \code{bounceFactor()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceFactor(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RFactor)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RFactor: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceFactorVector"></a>}}
	
	\subsection{Method \code{bounceFactorVector()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceFactorVector(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RFactorVector)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RFactorVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDate"></a>}}
	
	\subsection{Method \code{bounceDate()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceDate(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RDate)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RDate: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDateVector"></a>}}
	
	\subsection{Method \code{bounceDateVector()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceDateVector(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RDateVector)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RDateVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceLogical"></a>}}
	
	\subsection{Method \code{bounceLogical()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceLogical(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RLogical)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RLogical: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceLogicalVector"></a>}}
	
	\subsection{Method \code{bounceLogicalVector()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceLogicalVector(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RLogicalVector)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RLogicalVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDataframe"></a>}}
	
	\subsection{Method \code{bounceDataframe()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceDataframe(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RDataframe)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RDataframe: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceList"></a>}}
	
	\subsection{Method \code{bounceList()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceList(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RList)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RList: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNamedList"></a>}}
	
	\subsection{Method \code{bounceNamedList()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceNamedList(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RNamedList)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RNamedList: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceArray"></a>}}
	
	\subsection{Method \code{bounceArray()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$BounceTest$new();
instance$bounceArray(x)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{x}{ - (java expects a RNumericArray)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RNumericArray: 
		}

	}
}