% Generated by r6-generator-maven-plugin: do not edit by hand

\name{concat}
\alias{concat}
\title{no title}
\usage{concat(
	message1,
	message2
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{message1}{message1 - (java expects a RCharacter)}
	\item{message2}{message2 - (java expects a RCharacter)}
	\if{html}{\out{</div>}}
}
\value{
RCharacter: 
}
\description{
no description
}
\keyword{java api}


