% Generated by r6-generator-maven-plugin: do not edit by hand

\name{create}
\alias{create}
\title{A static object constructor}
\usage{create(
	message1,
	message2
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{message1}{message1 - the message to be printed - (java expects a RCharacter)}
	\item{message2}{message2 - will be used for toString - (java expects a RCharacter)}
	\if{html}{\out{</div>}}
}
\value{
R6 MoreFeatureTest object: A MoreFeatureTest R6 object
}
\description{
no description
}
\examples{
}
\keyword{java api}


