% Generated by r6-generator-maven-plugin: do not edit by hand

\name{demo_static}
\alias{demo_static}
\title{Static methods are also supported. }
\usage{demo_static(
	message
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{message}{message a message - (java expects a String)}
	\if{html}{\out{</div>}}
}
\value{
void: 
}
\description{
These are accessed through the
root of the R api, and as a functional interface
}
\examples{
}
\keyword{java api}


