% Generated by r6-generator-maven-plugin: do not edit by hand

\name{deserialise_list}
\alias{deserialise_list}
\title{no description}
\usage{
deserialise_list(
	filename
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{filename}{filename - (java expects a String)}
	\if{html}{\out{</div>}}
}
\value{
RList: 
}
\description{

}
\examples{\dontrun{
}
}
\keyword{java api}


