% Generated by r6-generator-maven-plugin: do not edit by hand

\name{FactoryTest}

\alias{FactoryTest}

\title{FactoryTest}



\description{
missing description

Version: 0.5.2

Generated: 2022-08-15T15:23:02.178
}

\details{
	no details
}

\examples{
## -----------------------------------
## Construct new instance of FactoryTest
## -----------------------------------
\dontrun{
J = testRapi::JavaApi$get()
# appropriate parameter values must be provided
instance = J$FactoryTest$new()
}

## -----------------------------------
## Method `FactoryTest$generateCharacter()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateCharacter()
}

## -----------------------------------
## Method `FactoryTest$generateNumeric()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateNumeric()
}

## -----------------------------------
## Method `FactoryTest$generateInteger()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateInteger()
}

## -----------------------------------
## Method `FactoryTest$generateFactor()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateFactor()
}

## -----------------------------------
## Method `FactoryTest$generateLogical()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateLogical()
}

## -----------------------------------
## Method `FactoryTest$generateCharacterVec()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateCharacterVec()
}

## -----------------------------------
## Method `FactoryTest$generateNumericVec()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateNumericVec()
}

## -----------------------------------
## Method `FactoryTest$generateIntegerVec()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateIntegerVec()
}

## -----------------------------------
## Method `FactoryTest$generateFactorVec()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateFactorVec()
}

## -----------------------------------
## Method `FactoryTest$generateLogicalVec()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateLogicalVec()
}

## -----------------------------------
## Method `FactoryTest$generateDataframe()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateDataframe()
}

## -----------------------------------
## Method `FactoryTest$generateStreamDataframe()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateStreamDataframe()
}

## -----------------------------------
## Method `FactoryTest$generateList()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateList()
}

## -----------------------------------
## Method `FactoryTest$generateNamedList()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$generateNamedList()
}
}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$FactoryTest$new()}}
		}
	}
	\subsection{Class methods}{
		\itemize{
			\item \href{#method-generateCharacter}{\code{instance$generateCharacter()}}
			\item \href{#method-generateNumeric}{\code{instance$generateNumeric()}}
			\item \href{#method-generateInteger}{\code{instance$generateInteger()}}
			\item \href{#method-generateFactor}{\code{instance$generateFactor()}}
			\item \href{#method-generateLogical}{\code{instance$generateLogical()}}
			\item \href{#method-generateCharacterVec}{\code{instance$generateCharacterVec()}}
			\item \href{#method-generateNumericVec}{\code{instance$generateNumericVec()}}
			\item \href{#method-generateIntegerVec}{\code{instance$generateIntegerVec()}}
			\item \href{#method-generateFactorVec}{\code{instance$generateFactorVec()}}
			\item \href{#method-generateLogicalVec}{\code{instance$generateLogicalVec()}}
			\item \href{#method-generateDataframe}{\code{instance$generateDataframe()}}
			\item \href{#method-generateStreamDataframe}{\code{instance$generateStreamDataframe()}}
			\item \href{#method-generateList}{\code{instance$generateList()}}
			\item \href{#method-generateNamedList}{\code{instance$generateNamedList()}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Constructor \code{new()}}{
the default no-args constructor

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
the new R6 FactoryTest object
		}

	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateCharacter"></a>}}
	
	\subsection{Method \code{generateCharacter()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateCharacter()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RCharacter: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateNumeric"></a>}}
	
	\subsection{Method \code{generateNumeric()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateNumeric()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RNumeric: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateInteger"></a>}}
	
	\subsection{Method \code{generateInteger()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateInteger()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RInteger: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateFactor"></a>}}
	
	\subsection{Method \code{generateFactor()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateFactor()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RFactor: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateLogical"></a>}}
	
	\subsection{Method \code{generateLogical()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateLogical()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RLogical: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateCharacterVec"></a>}}
	
	\subsection{Method \code{generateCharacterVec()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateCharacterVec()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RCharacterVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateNumericVec"></a>}}
	
	\subsection{Method \code{generateNumericVec()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateNumericVec()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RNumericVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateIntegerVec"></a>}}
	
	\subsection{Method \code{generateIntegerVec()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateIntegerVec()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RIntegerVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateFactorVec"></a>}}
	
	\subsection{Method \code{generateFactorVec()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateFactorVec()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RFactorVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateLogicalVec"></a>}}
	
	\subsection{Method \code{generateLogicalVec()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateLogicalVec()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RLogicalVector: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateDataframe"></a>}}
	
	\subsection{Method \code{generateDataframe()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateDataframe()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RDataframe: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateStreamDataframe"></a>}}
	
	\subsection{Method \code{generateStreamDataframe()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateStreamDataframe()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RDataframe: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateList"></a>}}
	
	\subsection{Method \code{generateList()}}{
Lists are much harder to type check than vectors hence RList builder methods throw checked exceptions

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateList()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RList: a RList containing the supplied Java objects converted into RObjects
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateNamedList"></a>}}
	
	\subsection{Method \code{generateNamedList()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new();
instance$generateNamedList()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RNamedList: 
		}

	}
}