% Generated by r6-generator-maven-plugin: do not edit by hand

\name{MinimalExample}

\alias{MinimalExample}

\title{MinimalExample}



\description{
This class is a very basic example of the features of the rJava maven plugin. 


Version: 0.5.2

Generated: 2022-08-15T15:23:02.281
}

\details{
	
The class is annotated with an @RClass to identify it as part of the R API. 

}

\examples{
## -----------------------------------
## Construct new instance of MinimalExample
## -----------------------------------
\dontrun{
J = testRapi::JavaApi$get()
# appropriate parameter values must be provided
instance = J$MinimalExample$new()
}

## -----------------------------------
## Method `MinimalExample$demo(dataframe, message)`
## -----------------------------------
J = JavaApi$get()
}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$MinimalExample$new()}}
		}
	}
	\subsection{Class methods}{
		\itemize{
			\item \href{#method-demo}{\code{instance$demo(dataframe, message)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Constructor \code{new()}}{
the default no-args constructor

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MinimalExample$new();
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
the new R6 MinimalExample object
		}

	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-demo"></a>}}
	
	\subsection{Method \code{demo()}}{
Documentation of the method can be done in JavaDoc and these will be present in the R documentation

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MinimalExample$new();
instance$demo(dataframe, message)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{dataframe - a dataframe with an arbitrary number of columns}{ - (java expects a RDataframe)}
				\item{message - a message}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RDataframe: the dataframe unchanged
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
J = JavaApi$get()
			}
			\if{html}{\out{</div>}}
		}
	}
}