% Generated by r6-generator-maven-plugin: do not edit by hand

\name{more_feature_test_collider}
\alias{more_feature_test_collider}
\title{no title}
\usage{more_feature_test_collider(
	message1,
	message2
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{message1}{message1 - (java expects a RCharacter)}
	\item{message2}{message2 - (java expects a RCharacter)}
	\if{html}{\out{</div>}}
}
\value{
RCharacter: 
}
\description{
no description
}
\examples{
library(testthat)
tmp = more_feature_test_collider('should ','work')
expect_equal(tmp,'more feature test: should work')
}
\keyword{java api}


