% Generated by r6-generator-maven-plugin: do not edit by hand

\name{MoreFeatureTest}

\alias{MoreFeatureTest}

\title{MoreFeatureTest}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{message1}{message1 - the message to be printed - (java expects a RCharacter)}
	\item{message2}{message2 - will be used for toString - (java expects a RCharacter)}
	\if{html}{\out{</div>}}
}


\description{
This has no documentation

Version: 0.6.1.9000

Generated: 2024-04-22T20:47:47.434003263
}

\details{
	This has no documentation
}

\examples{
## -----------------------------------
## Construct new instance of MoreFeatureTest
## -----------------------------------
\dontrun{
J = testRapi::JavaApi$get()
# appropriate parameter values must be provided
instance = J$MoreFeatureTest$new(message1, message2)
}

## -----------------------------------
## Method `MoreFeatureTest$testLogging()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$testLogging()
}

## -----------------------------------
## Method `MoreFeatureTest$throwCatchable()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$throwCatchable()
}

## -----------------------------------
## Method `MoreFeatureTest$printMessage()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$printMessage()
}

## -----------------------------------
## Method `MoreFeatureTest$throwRuntime()`
## -----------------------------------
\dontrun{
# appropriate parameter values must be provided
instance$throwRuntime()
}
}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$MoreFeatureTest$new(message1, message2)}}
		}
	}
	\subsection{Class methods}{
		\itemize{
			\item \href{#method-testLogging}{\code{instance$testLogging()}}
			\item \href{#method-throwCatchable}{\code{instance$throwCatchable()}}
			\item \href{#method-printMessage}{\code{instance$printMessage()}}
			\item \href{#method-throwRuntime}{\code{instance$throwRuntime()}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Constructor \code{new()}}{
the first constructor is used if there are none annotated

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MoreFeatureTest$new(message1, message2);
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{message1 - the message to be printed}{ - (java expects a RCharacter)}
				\item{message2 - will be used for toString}{ - (java expects a RCharacter)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
the new R6 MoreFeatureTest object
		}

	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-testLogging"></a>}}
	
	\subsection{Method \code{testLogging()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MoreFeatureTest$new(message1, message2);
instance$testLogging()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
void: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-throwCatchable"></a>}}
	
	\subsection{Method \code{throwCatchable()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MoreFeatureTest$new(message1, message2);
instance$throwCatchable()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RCharacter: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-printMessage"></a>}}
	
	\subsection{Method \code{printMessage()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MoreFeatureTest$new(message1, message2);
instance$printMessage()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
void: 
		}

	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-throwRuntime"></a>}}
	
	\subsection{Method \code{throwRuntime()}}{
no title

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MoreFeatureTest$new(message1, message2);
instance$throwRuntime()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
RCharacter: 
		}

	}
}