% Generated by r6-generator-maven-plugin: do not edit by hand

\name{serialise_dataframe}
\alias{serialise_dataframe}
\title{no title}
\usage{serialise_dataframe(
	dataframe,
	filename
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{dataframe}{dataframe - (java expects a RDataframe)}
	\item{filename}{filename - (java expects a String)}
	\if{html}{\out{</div>}}
}
\value{
void: 
}
\description{
no description
}
\keyword{java api}


