% Generated by r6-generator-maven-plugin: do not edit by hand

\name{Serialiser}

\alias{Serialiser}

\title{Serialiser}



\description{
Actually useful class which takes an R dataframe and a filename

Version: 1.1.0

Generated: 2024-05-16T16:22:15.497640740
}

\details{
	
string and serialises the dataframe so it can be used natively in java
for testing purposes.
}

\examples{
## -----------------------------------
## Construct new instance of Serialiser
## -----------------------------------
\dontrun{
J = testRapi::JavaApi$get()
# appropriate parameter values must be provided
instance = J$Serialiser$new()
}
}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$Serialiser$new()}}
		}
	}
	\subsection{Class methods}{
		\itemize{
			\item none
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Constructor \code{new()}}{
the default no-args constructor

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$Serialiser$new();
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item none
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
the new R6 Serialiser object
		}

	}

}