% Generated by r6-generator-maven-plugin: do not edit by hand

\name{deserialise_dataframe}
\alias{deserialise_dataframe}
\title{no title}
\usage{deserialise_dataframe(
	filename
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{filename}{filename - (java expects a String)}
	\if{html}{\out{</div>}}
}
\value{
RDataframe: 

}
\description{
no description
}
\keyword{java api}


