% Generated by r6-generator-maven-plugin: do not edit by hand

\name{diamonds}
\alias{diamonds}
\title{The ggplot2::diamonds dataframe }
\usage{diamonds()
}
\value{
RDataframe: the ggplot2::diamonds dataframe
}
\description{
A copy serialised into java, using
RObject.writeRDS, saved within the jar file of the package, and exposed here
using RObject.readRDS.
}
\examples{
diamonds()
}
\keyword{java api}


