% Generated by r6-generator-maven-plugin: do not edit by hand

\name{serialise_list}
\alias{serialise_list}
\title{no description}
\usage{
serialise_list(
	dataframe,
	filename
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{dataframe}{dataframe - (java expects a RList)}
	\item{filename}{filename - (java expects a String)}
	\if{html}{\out{</div>}}
}
\value{
void: 
}
\description{

}
\examples{\dontrun{
}
}
\keyword{java api}


