% Generated by r6-generator-maven-plugin: do not edit by hand
\docType{package}
\name{testRapi-package}
\alias{testRapi}
\alias{testRapi-package}
\title{testRapi: A Test Library}
\description{
% \if{html}{\figure{logo.png}{options: align='right'}}
Documents the features of the 'r6-generator-maven-plugin' by providing an example of an R package automatically generated from Java code by the plugin. It is not intended to be useful beyond testing, demonstrating and documenting the features of the plugin.

Version: 0.5.6

Classes:
\itemize{ 
	\item \code{JavaApi}
	\item \code{MoreFeatureTest}
	\item \code{Serialiser}
	\item \code{FeatureTest}
	\item \code{BounceTest}
	\item \code{FactoryTest}
	\item \code{MinimalExample}
}
}

\author{
\itemize{ 
   \item Rob Challen rc538@exeter.ac.uk 
   \item terminological   
   \item Rob Challen rob@terminological.co.uk 0000-0002-5504-7768
}
}

\keyword{java api}